fix warning ("left shift of negative value") using a cast

+++ src/dsp/lossless_enc_sse2.c
@@ -19,7 +19,7 @@
 #include "./lossless.h"
 
 // For sign-extended multiplying constants, pre-shifted by 5:
-#define CST_5b(X)  (((int16_t)(X << 8)) >> 5)
+#define CST_5b(X)  (((int16_t)((uint16_t)X << 8)) >> 5)
 
 //------------------------------------------------------------------------------
 // Subtract-Green Transform
